Fix test-server.
